% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.na_result}
\alias{.na_result}
\title{Uniform negative-result list returned when inputs are nonsensical.}
\usage{
.na_result(message)
}
\description{
Uniform negative-result list returned when inputs are nonsensical.
}
