% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{gene_annotation_tools}
\alias{gene_annotation_tools}
\title{All nine gene annotation MCP tools as a list}
\format{
An object of class \code{list} of length 9.
}
\usage{
gene_annotation_tools
}
\description{
A named list of all \code{\link[ellmer:tool]{ellmer::tool()}} objects defined in this package,
ready to pass to \code{mcptools::mcp_server(tools = gene_annotation_tools)}.
}
\keyword{datasets}
